# oas/swagger.yml
openapi: 3.0.1
info:
  title: QA API
  description: PactFlow QA API demo
  version: 1.0.0
paths:
  /QAs:
    post:
      summary: Create a QA
      description: Creates a new QA
      operationId: createQA
      requestBody:
        description: Create a new QA
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QA'
            examples:
              application/json:
                value:
                  id: "1"
                  name: "paddy"
                  age: 42
      responses:
        "200":
          description: successful operation
          content:
            "application/json; charset=utf-8":
              schema:
                $ref: '#/components/schemas/QA'
              examples:
                application/json:
                  value:
                    id: "1234"
                    name: "paddy"
                    age: 42
    get:
      summary: List all QAs
      description: Returns all QAs
      operationId: getAllQAs
      responses:
        "200":
          description: successful operation
          content:
            "application/json; charset=utf-8":
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/QA'
              examples:
                application/json:
                  value:
                    - id: "1234"
                      name: "food"
                      age: 42
                      # name: "pizza"
                      # version: "1.0.0"
                      # see https://github.com/apiaryio/dredd/issues/1430 for why
        "400":
          description: Invalid ID supplied
          content: {}
  /QA/{id}:
    get:
      summary: Find QA by ID
      description: Returns a single QA
      operationId: getQAByID
      parameters:
        - name: id
          in: path
          description: ID of QA to get
          schema:
            type: string
          required: true
          example: 10
      responses:
        "200":
          description: successful operation
          content:
            "application/json; charset=utf-8":
              schema:
                $ref: '#/components/schemas/QA'
              examples:
                application/json:
                  value:
                    id: "1234"
                    name: "food"
                    age: 42
                    # name: "pizza"
                    # version: "1.0.0"
                    # see https://github.com/apiaryio/dredd/issues/1430 for why
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: QA not found
          content: {}
components:
  schemas:
    QA:
      type: object
      required:
        - id
        - name
        - age
      properties:
        id:
          type: string
        name:
          type: string
        surname:
          type: string
        version:
          type: string
        age:
          type: number